Setup first of many new test files
Taking a new approach to testing... each file is a module. Each module
covers a single model that represents a particular type of scenario the
app needs to handle. Every module should have the same number of
tests... one for each piece of functionality of the frame work. So every
module has complete test coverage.