Added a more complete type checking solution
This solution is a 90% solution for cross-frame type checking. Defaults
to the

Object.prototype.toString.call(obj)

method if the crossFrameSafe boolean is set to true. This adds quite a
bit of bulk to accomplish cross frame safety, and might not be worth it.

According to research here-

http://whattheheadsaid.com/2010/10/cross-context-isarray-and-internet-explorer

IE versions earlier than 9, in a cross-frame scenario, toString will
always resolve to [object Object].